Moved getDirectiveLocations
